Fixing if condition bug!
The right side of an `||` expression always evaluated to false because it was testing that the same byte is simultaneously equal to four different values, which is logically impossible without being `volatile` (and if even if it were `volatile` it would almost certainly still be a bug).

Instead, I believe the intention was to work like the left-hand expression, which tests four consecutive bytes. I've changed it to do that.